Домашнее задание №18-19
