+++ app/src/main/AndroidManifest.xml
@@ -13,63 +13,63 @@
         android:theme="@style/Theme.Homework17"
         tools:targetApi="31">
         <activity
-            android:name=".AddNotesActivity"
+            android:name=".activity.notesActivity.AddNotesActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".ListOfNotesActivity"
+            android:name=".activity.notesActivity.ListOfNotesActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".LogInActivity"
+            android:name=".activity.authorizeActivity.LogInActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".SignUpActivity"
+            android:name=".activity.authorizeActivity.SignUpActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".FifthDiscoverActivity"
+            android:name=".activity.discoverActivity.FifthDiscoverActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".FourthDiscoverActivity"
+            android:name=".activity.discoverActivity.FourthDiscoverActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".ThirdDiscoverActivity"
+            android:name=".activity.discoverActivity.ThirdDiscoverActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".SecondDiscoverActivity"
+            android:name=".activity.discoverActivity.SecondDiscoverActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".FirstDiscoverActivity"
+            android:name=".activity.discoverActivity.FirstDiscoverActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
