+++ app/src/main/java/com/example/homework17/activity/authorizeActivity/LogInActivity.kt
@@ -1,12 +1,14 @@
-package com.example.homework17
+package com.example.homework17.activity.authorizeActivity
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.Gravity
 import android.widget.TextView
-import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.AppCompatButton
+import com.example.homework17.R
+import com.example.homework17.activity.notesActivity.ListOfNotesActivity
+import com.example.homework17.util.TextFieldUtil
+import com.example.homework17.util.ToastUtil
 import com.google.android.material.textfield.TextInputLayout
 
 class LogInActivity : AppCompatActivity() {
@@ -14,6 +16,8 @@ class LogInActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_log_in)
+        val toast = ToastUtil()
+        val text = TextFieldUtil()
 
         val buttonLogIn: AppCompatButton = findViewById(R.id.buttonLogIn)
 
@@ -24,13 +28,13 @@ class LogInActivity : AppCompatActivity() {
 
 
         buttonLogIn.setOnClickListener {
-            val isEmailValid = checker(emailFieldLayout, R.string.error_blank)
-            val isPasswordValid = checker(passwordFieldLayout, R.string.error_blank)
+            val isEmailValid = text.isFieldBlank(emailFieldLayout, R.string.error_blank)
+            val isPasswordValid = text.isFieldBlank(passwordFieldLayout, R.string.error_blank)
 
             if (isEmailValid && isPasswordValid) {
                 startActivity(intentNoteList)
             } else {
-                toast(R.string.toast_error)
+                toast.create(R.string.toast_error)
             }
 
         }
@@ -42,20 +46,5 @@ class LogInActivity : AppCompatActivity() {
         }
     }
 
-    private fun checker(fieldLayout: TextInputLayout, error: Int): Boolean {
-        return if (fieldLayout.editText?.text.isNullOrBlank()) {
-            fieldLayout.error = getString(error)
-            false
-        } else {
-            fieldLayout.error = null
-            true
-        }
-    }
 
-    private fun toast(toastMessage: Int) {
-        val toastText = getString(toastMessage)
-        val toast = Toast.makeText(applicationContext, toastText, toastText.length)
-        toast.setGravity(Gravity.CENTER, 0, 0)
-        toast.show()
-    }
 }
