+++ app/src/main/java/com/example/homework17/activity/discoverActivity/FifthDiscoverActivity.kt
@@ -1,19 +1,20 @@
-package com.example.homework17
+package com.example.homework17.activity.discoverActivity
 
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
 import android.widget.TextView
+import androidx.appcompat.app.AppCompatActivity
 import androidx.core.os.postDelayed
+import com.example.homework17.R
+import com.example.homework17.activity.authorizeActivity.SignUpActivity
+import com.example.homework17.handler
 
-class ThirdDiscoverActivity : AppCompatActivity() {
+class FifthDiscoverActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_third_discover)
+        setContentView(R.layout.activity_fifth_discover)
 
-        val intentNextActivity = Intent(this, FourthDiscoverActivity::class.java)
+        val intentNextActivity = Intent(this, SignUpActivity::class.java)
         val buttonSkip: TextView = findViewById(R.id.skip)
 
         handler.postDelayed(10000) {
