+++ app/src/main/java/com/example/homework17/activity/discoverActivity/FirstDiscoverActivity.kt
@@ -1,10 +1,12 @@
-package com.example.homework17
+package com.example.homework17.activity.discoverActivity
 
 import android.content.Intent
 import android.os.Bundle
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.os.postDelayed
+import com.example.homework17.R
+import com.example.homework17.handler
 
 class FirstDiscoverActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
