+++ app/src/main/res/layout/activity_add_notes.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/background"
     android:orientation="vertical"
-    tools:context=".AddNotesActivity">
+    tools:context=".activity.notesActivity.AddNotesActivity">
 
     <TextView
         android:id="@+id/back"
