+++ app/src/main/res/layout/activity_first_discover.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bubbles_1_discover"
-    tools:context=".FirstDiscoverActivity">
+    tools:context=".activity.discoverActivity.FirstDiscoverActivity">
 
     <TextView
         android:id="@+id/skip"
