+++ app/src/main/res/layout/activity_fourth_discover.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bubbles__4_discover"
-    tools:context=".FourthDiscoverActivity">
+    tools:context=".activity.discoverActivity.FourthDiscoverActivity">
 
     <TextView
         android:id="@+id/skip"
