+++ app/src/main/res/layout/activity_list_of_notes.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ListOfNotesActivity"
+    tools:context=".activity.notesActivity.ListOfNotesActivity"
     android:orientation="vertical"
     android:background="@color/background">
