+++ app/src/main/res/layout/activity_second_discover.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SecondDiscoverActivity">
+    tools:context=".activity.discoverActivity.SecondDiscoverActivity">
 
     <FrameLayout
         android:layout_width="match_parent"
