+++ app/src/main/res/layout/activity_sign_up.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bubbles_sign_up"
-    tools:context=".SignUpActivity">
+    tools:context=".activity.authorizeActivity.SignUpActivity">
 
     <LinearLayout
         android:layout_width="@dimen/dp343"
