+++ app/src/main/res/layout/activity_third_discover.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bubbles__3_discover"
-    tools:context=".ThirdDiscoverActivity">
+    tools:context=".activity.discoverActivity.ThirdDiscoverActivity">
 
     <TextView
         android:id="@+id/skip"
