+++ app/src/main/res/layout/item_note.xml
@@ -1,20 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    android:layout_margin="10dp"
     android:background="@color/background"
-    android:layout_margin="10dp">
+    android:orientation="vertical">
+
     <TextView
         android:id="@+id/title"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:textSize="20sp"
-        android:fontFamily="@font/red_hat_display"
-        android:textColor="@color/title"
         android:layout_gravity="top|start"
         android:layout_marginStart="16dp"
-        />
+        android:fontFamily="@font/red_hat_display"
+        android:hint="@string/title"
+        android:textColor="@color/title"
+        android:textSize="20sp" />
 
     <TextView
         android:id="@+id/date"
@@ -22,18 +23,15 @@ android:orientation="vertical"
         android:layout_height="wrap_content"
         android:layout_gravity="end|center_horizontal"
         android:layout_marginEnd="@dimen/dp50"
+        android:hint="@string/date"
         android:textColor="@color/purple_hint"
         android:textSize="13sp" />
 
     <TextView
         android:id="@+id/message"
-        android:layout_marginTop="@dimen/dp50"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:textSize="16sp"
-        />
-
-
-
-
+        android:layout_marginTop="@dimen/dp50"
+        android:hint="@string/message"
+        android:textSize="16sp" />
 </FrameLayout>
