+++ app/src/main/res/values/strings.xml
@@ -44,6 +44,7 @@
     <string name="title">Title</string>
     <string name="message">Message</string>
     <string name="add">Add</string>
+    <string name="date">date</string>
 
 
 </resources>
